refactored generic subgraph generator
