+++ src/rust/generic-subgraph-generator/src/models/file/mod.rs
@@ -1,0 +1,9 @@
+mod create;
+mod delete;
+mod write;
+mod read;
+
+pub use create::FileCreate;
+pub use delete::FileDelete;
+pub use read::FileRead;
+pub use write::FileWrite;
