+++ src/rust/generic-subgraph-generator/src/models/network/mod.rs
@@ -1,0 +1,5 @@
+mod outbound;
+mod inbound;
+
+pub use outbound::ProcessOutboundConnectionLog;
+pub use inbound::ProcessInboundConnectionLog;
